+++ package.json
@@ -40,8 +40,8 @@
     "@coolwallets/wallet": "0.2.0-beta.1",
     "@ensdomains/dnsregistrar": "0.4.0",
     "@myetherwallet/mew-components": "0.5.32-beta",
-    "@sentry/browser": "6.1.0",
-    "@sentry/integrations": "6.1.0",
+    "@sentry/vue": "6.2.0",
+    "@sentry/tracing": "6.2.0",
     "@unstoppabledomains/resolution": "1.17.0",
     "@vue/babel-preset-app": "4.5.11",
     "@walletconnect/browser": "1.0.0",
@@ -56,7 +56,7 @@
     "apollo-utilities": "1.3.4",
     "bitbox02-api": "0.8.0",
     "clipboard-copy": "4.0.1",
-    "core-js": "3.8.3",
+    "core-js": "3.9.0",
     "graphql": "15.5.0",
     "graphql-anywhere": "4.2.7",
     "graphql-tag": "2.11.0",
@@ -131,7 +131,7 @@
     "remark-cli": "9.0.0",
     "remark-preset-lint-recommended": "5.0.0",
     "rskjs-util": "1.0.3",
-    "sass": "1.32.7",
+    "sass": "1.32.8",
     "sass-loader": "10.1.1",
     "sinon": "9.2.4",
     "store": "2.0.12",
@@ -165,8 +165,7 @@
     "axios": "^0.21.1",
     "node-fetch": "^2.6.1",
     "elliptic": "^6.5.3",
-    "cryptiles": "^4.1.2",
-    "bl": "^1.2.3"
+    "cryptiles": "^4.1.2"
   },
   "license": "MIT",
   "repository": "https://github.com/myetherwallet/myetherwallet"
