+++ src/main/components.js
@@ -1,0 +1,8 @@
+import Vue from 'vue';
+import mewComponents from '@myetherwallet/mew-components';
+import whiteSheet from '@/components/white-sheet/WhiteSheet.vue';
+
+Object.keys(mewComponents).forEach(name => {
+  Vue.component(name, mewComponents[name]);
+});
+Vue.component('Mew6WhiteSheet', whiteSheet);
