--- src/utils/partners/bity/index.js
@@ -1,4 +1,0 @@
-import BitySwap from './bity';
-import { BityCurrencies, PROVIDER_NAME } from './config';
-
-export { BitySwap, BityCurrencies, PROVIDER_NAME };
