--- src/utils/partners/changelly/config/index.js
@@ -1,21 +1,0 @@
-import { ChangellyCurrencies } from './currencies';
-import {
-  changellyMethods,
-  statuses,
-  notificationStatuses,
-  TIME_SWAP_VALID,
-  requireExtraId,
-  PROVIDER_NAME,
-  FEE_RATE
-} from './config';
-
-export {
-  ChangellyCurrencies,
-  changellyMethods,
-  statuses,
-  notificationStatuses,
-  TIME_SWAP_VALID,
-  requireExtraId,
-  PROVIDER_NAME,
-  FEE_RATE
-};
