--- src/utils/partners/changelly/index.js
@@ -1,4 +1,0 @@
-import { ChangellyCurrencies, PROVIDER_NAME } from './config';
-import ChangellySwap from './changelly';
-
-export { ChangellySwap, ChangellyCurrencies, PROVIDER_NAME };
