--- src/utils/partners/helpers/haveIcon.js
@@ -1,12 +1,0 @@
-import { haveIcon } from '../partnersConfig';
-
-const hasIcon = currency => {
-  if (haveIcon.includes(currency)) {
-    return currency;
-  } else if (haveIcon.includes(`${currency}-alt`)) {
-    return `${currency}-alt`;
-  }
-  return '';
-};
-
-export default hasIcon;
