--- src/utils/partners/kyber/config/currenciesKOV.js
@@ -1,70 +1,0 @@
-const KyberCurrenciesKOV = {
-  ETH: {
-    symbol: 'ETH',
-    name: 'Ethereum',
-    address: '0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee',
-    decimals: 18
-  },
-  KNC: {
-    symbol: 'KNC',
-    name: 'Kyber Network',
-    address: '0xad67cB4d63C9da94AcA37fDF2761AaDF780ff4a2',
-    decimals: 18
-  },
-  OMG: {
-    symbol: 'OMG',
-    name: 'OmiseGO',
-    address: '0xdB7ec4E4784118D9733710e46F7C83fE7889596a',
-    decimals: 18
-  },
-  DAI: {
-    symbol: 'DAI',
-    name: 'Dai',
-    address: '0xC4375B7De8af5a38a93548eb8453a498222C4fF2',
-    decimals: 18
-  },
-  MKR: {
-    symbol: 'MKR',
-    name: 'Maker',
-    address: '0xAaF64BFCC32d0F15873a02163e7E500671a4ffcD',
-    decimals: 18
-  },
-  POLY: {
-    symbol: 'POLY',
-    name: 'Polymath',
-    address: '0xd92266fd053161115163a7311067F0A4745070b5',
-    decimals: 18
-  },
-  MANA: {
-    symbol: 'MANA',
-    name: 'Decentraland',
-    address: '0xf5E314c435B3B2EE7c14eA96fCB3307C3a3Ef608',
-    decimals: 18
-  },
-  SALT: {
-    symbol: 'SALT',
-    name: 'Salt',
-    address: '0x6fEE5727EE4CdCBD91f3A873ef2966dF31713A04',
-    decimals: 18
-  },
-  SNT: {
-    symbol: 'SNT',
-    name: 'Status Network',
-    address: '0x4c99B04682fbF9020Fcb31677F8D8d66832d3322',
-    decimals: 18
-  },
-  WETH: {
-    symbol: 'WETH',
-    name: 'Wrapped Ether',
-    address: '0xd0A1E359811322d97991E03f863a0C30C2cF029C',
-    decimals: 18
-  },
-  ZIL: {
-    symbol: 'ZIL',
-    name: 'Zilliqa',
-    address: '0xAb74653cac23301066ABa8eba62b9Abd8a8c51d6',
-    decimals: 18
-  }
-};
-
-export { KyberCurrenciesKOV };
