--- src/utils/partners/kyber/index.js
@@ -1,4 +1,0 @@
-import KyberSwap from './kyber';
-import { KyberCurrencies, PROVIDER_NAME } from './config';
-
-export { KyberSwap, KyberCurrencies, PROVIDER_NAME };
