--- src/utils/partners/partnersConfig/index.js
@@ -1,34 +1,0 @@
-import { chainCurrencies, fiat } from './currencyDetails';
-import EthereumTokens from './EthereumTokens';
-import OtherCoins from './OtherCoins';
-import ERC20 from './ERC20Token.abi';
-import haveIcon from './haveIcon';
-import {
-  networkSymbols,
-  TIME_SWAP_VALID_DEFAULT,
-  BASE_CURRENCY,
-  MIN_SWAP_AMOUNT,
-  TOP_OPTIONS_ORDER,
-  swapNotificationStatuses,
-  baseCurrencyEntry,
-  swapApiEndpoints,
-  mewSimplex
-} from './config';
-
-export {
-  ERC20,
-  BASE_CURRENCY,
-  MIN_SWAP_AMOUNT,
-  TIME_SWAP_VALID_DEFAULT,
-  TOP_OPTIONS_ORDER,
-  swapNotificationStatuses,
-  baseCurrencyEntry,
-  EthereumTokens,
-  OtherCoins,
-  haveIcon,
-  networkSymbols,
-  chainCurrencies,
-  fiat,
-  swapApiEndpoints,
-  mewSimplex
-};
