--- src/utils/partners/simplex/index.js
@@ -1,4 +1,0 @@
-import Simplex from './simplex';
-import { SimplexCurrencies, PROVIDER_NAME } from './config';
-
-export { Simplex, SimplexCurrencies, PROVIDER_NAME };
